+++ cpp/mesh_gen.cpp
@@ -9,18 +9,18 @@ Vec3<double> spherify(const Vec3<double>& center, double r, const Vec3<double>& 
     return (r / dist(x, center)) * (x - center) + center;
 }
 
-Mesh<3> sphere_mesh(const Vec3<double>& center, double r, int refinements) {
+Mesh<3> sphere_mesh(const Vec3<double>& center, double r, size_t refinements) {
     std::vector<Vec3<double>> vertices =
     {
         {0.0, -r, 0.0}, {r, 0.0, 0.0}, {0.0, 0.0, r},
         {-r, 0.0, 0.0}, {0.0, 0.0, -r}, {0.0, r, 0.0}
     };
 
-    for (unsigned int i = 0; i < vertices.size(); i++) {
+    for (size_t i = 0; i < vertices.size(); i++) {
         vertices[i] += center;
     }
 
-    std::vector<std::array<int,3>> faces = 
+    std::vector<std::array<size_t,3>> faces = 
     {
         {1, 0, 2}, {2, 0, 3}, {3, 0, 4}, {4, 0, 1},
         {5, 1, 2}, {5, 2, 3}, {5, 3, 4}, {5, 4, 1}
@@ -46,7 +46,7 @@ Mesh<3> rect_mesh(const Vec3<double>& lower_left, const Vec3<double>& upper_left
     std::vector<Vec3<double>> vertices = {
         lower_left, upper_left, upper_right, lower_right
     };
-    std::vector<std::array<int,3>> faces = {
+    std::vector<std::array<size_t,3>> faces = {
         {0, 3, 2}, {0, 2, 1}
     };
     return Mesh<3>::from_vertices_faces(vertices, faces);
@@ -54,7 +54,7 @@ Mesh<3> rect_mesh(const Vec3<double>& lower_left, const Vec3<double>& upper_left
 
 Mesh<2> line_mesh(const Vec2<double>& a, const Vec2<double>& b) {
     std::vector<Vec<double,2>> vertices = {a, b};
-    std::vector<std::array<int,2>> segs = {{0, 1}};
+    std::vector<std::array<size_t,2>> segs = {{0, 1}};
     return Mesh<2>::from_vertices_faces(vertices, segs);
 }
 
@@ -63,11 +63,11 @@ Vec2<double> circlify(const Vec2<double>& center, double r, const Vec2<double>& 
     return (r / dist(x, center)) * (x - center) + center;
 }
 
-Mesh<2> circle_mesh(const Vec2<double>& c, double r, int refinements) {
+Mesh<2> circle_mesh(const Vec2<double>& c, double r, size_t refinements) {
     std::vector<std::array<double,2>> vertices = {
         {c[0] + r, c[1]}, {c[0], c[1] + r}, {c[0] - r, c[1]}, {c[0], c[1] - r},
     };
-    std::vector<std::array<int, 2>> segs = {
+    std::vector<std::array<size_t, 2>> segs = {
         {0, 1}, {1, 2}, {2, 3}, {3, 0}
     };
